Extract scala 2.11 specific tests; Fix build for 2.10
