+++ src/test/scala/com/softwaremill/macros/customize/ToStringMaskTest.scala
@@ -1,6 +1,7 @@
+package com.softwaremill.macros.customize
+
 import java.time.ZonedDateTime
 
-import com.softwaremill.macros.customize.{customize, mask}
 import org.scalatest.{FlatSpec, Matchers}
 
 class ToStringMaskTest extends FlatSpec with Matchers {
@@ -58,14 +59,6 @@ class ToStringMaskTest extends FlatSpec with Matchers {
     sc.toString should be("SmallClass(***)")
   }
 
-  it should "work on classes with > 22 fields" in {
-    // given
-    val bc = BigClass(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13 ,14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25)
-
-    // then
-    bc.toString should be("BigClass(***,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25)")
-  }
-
 }
 
 @customize
@@ -82,11 +75,6 @@ case class EmptyClass()
 
 @customize
 case class SmallClass(@mask n: Int)
-
-@customize
-case class BigClass(@mask a1: Int, a2: Int, a3: Int, a4: Int, a5: Int, a6: Int, a7: Int, a8: Int, a9: Int, a10: Int,
-                    a11: Int, a12: Int, a13: Int, a14: Int,a15: Int, a16: Int, a17: Int, a18: Int, a19: Int, a20: Int,
-                    a21: Int, a22: Int, a23: Int, a24: Int, a25: Int)
 
 object UserPrivate {
   def apply(email: String) = new UserPrivate(1, email)
